Be backwards compatible with the Windows Meterpreter
The older Windows Meterpreter had a bug where the core command IDs would
not be properly returned and added to the command array. Until the major
version of the payloads gem is bumped to ensure that all instances have
that change, skip validating Windows core command IDs. Any commands that
skip the proactive validation and fail, will continue to fail as they
did before.