guard for removed workspace
By caching the `current` workspace in the dataservice there is
introduced a possible case where the cached Rails entity may be
removed from the database by another console instance.

There are many possible edge cases for this however for now
if the workspace is missing fallback to existing guard behavior
and set the current cached workspace to be the `default` workspace.