Delineate Rex::Proto::*::Server mixins for Msf
Msf::Exploit::Remote::*::*Server mixins are implemented either as
descendants of the SocketServer mixin providing ports as service
objects and handling protocol interactions using handlers on those,
or as actual *Server instances out of Rex::Proto. When creating
MetasploitModules (or other namespaces) which mix two or more of
these Msf::Exploit::Remote::*::*Server mixins together, a namespace
collision occurs on the :service object of the namespace mixing in
the two setters for it and creating a last-write-wins condition on
self.server which causes the first server instantiated to be lost
into ObjectSpace - no rational reference to it, and no GC on it as
it has its own references to unmanaged memory (file descriptors).

Address the problem by delineating each of the Rex::Proto servers
with named prefixes such as http_service and dns_service in order
to permit mixing them all together along with the socket services
which can run arbitrary protocol handlers from the Msf namespace.
Leave backward compatibility via :service decorator methods in the
updated Msf Server mixins and method aliases to their start/stop
methods for access when overriding names are mixed in after these.