TTY acqusition for shell sessions
There are times in life when you need a tty - such as when `sudo`
asks you to input that credential you worked so hard to get.

Aquire said TTY using expect or a Python PTY spawn, each with its
benefits and downsides. The expect approach drops a tempfile, the
python approach can linger on the system in the process table.

This module was massaged and cleaned up for release: could benefit
from some fresh eyes on it.

Testing:
  The original was written in 2012, i somehow messed up getting
it upstream into the Framework. It has been used on a wide array
of targets/platforms with significant success - `sudo -s` with a
password input is possible with this.