vcenter_secrets_dump external PSC mod
Various tweaks and bug fixes following additional real-world testing.

Tweaked the SQL query for gathering ESXi hostnames to sort by host FQDN
for readability.

Add code to detect an external vSphere PSC in the topology, and report
this in the status banner. Previously the module blindly assumed it was
running against embedded, and it currently cannot handle the case where
vCenter and PSC are separate if execution is attempted from the PSC vs.
vCenter.

The module can now cope with separate vCenter management and PSC
appliances if it is executed on the vCenter management appliance; only
secrets germane to vCenter management will be extracted, such as ESXi
vpxuser passwords and secrets residing in customizations specs. VMCA
root, IdP certs, tenant AES / SSO identity sources will not be extracted
in this scenario.

Changed the code for grabbing the hostname to use VAMI instead of the
built-in hostname command to improve reliability of embedded PSC
detection.

Updated docs.